+++ src/main/java/org/javlo/ecom/coda/CodaFormatError.java
@@ -1,0 +1,5 @@
+package org.javlo.ecom.coda;
+
+public class CodaFormatError extends Exception {
+
+}